# api/kontablo_openapi.yaml
openapi: "3.1.0"
info:
  title: Kontablo Universal Accounting API
  description: |
    The Kontablo API provides universal accounting account mapping, transaction
    classification, and cross-jurisdiction financial consolidation capabilities.

    ## Core Capabilities
    - **Account Mapping**: Map local ERP account codes to the universal Kontablo ontology
    - **Transaction Classification**: Classify raw transactions (journal entries) to Kontablo accounts  
    - **Consolidation**: Aggregate multi-jurisdiction financial data into a unified Kontablo statement
    - **Validation**: Validate financial statement integrity (balance sheet equation, IAS compliance)
    - **Synonyms**: Query the multilingual accounting synonym corpus

    ## Authentication
    All endpoints require a Bearer JWT token obtained from `/auth/token`.

    ## Rate Limits
    - Free tier: 100 requests/day
    - Pro tier: 10,000 requests/day
    - Enterprise: unlimited

    ## Versioning
    This is API version `v1`. Breaking changes will increment to `v2`.

  version: "1.0.0"
  contact:
    name: Kontablo Open Standard
    url: https://github.com/ChristianLuciani/accounting-esperanto
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kontablo.io/v1
    description: Production
  - url: https://staging-api.kontablo.io/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /accounts:
    get:
      tags: [accounts]
      summary: List all Kontablo accounts
      operationId: accountsList
      parameters:
        - name: level
          in: query
          schema:
            type: integer
            enum: [1, 2, 3]
          description: Filter by ontology level (1=top-level, 3=operational)
        - name: nature
          in: query
          schema:
            type: string
            enum: [debit, credit]
        - name: statement
          in: query
          schema:
            type: string
            enum: [balance_sheet, income_statement, cash_flow]
        - name: parent
          in: query
          schema:
            type: string
          description: Filter by parent account ID (e.g., "asset.current")
      responses:
        "200":
          description: List of Kontablo accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/KontabloAccount"
  /accounts/{account_id}:
    get:
      tags: [accounts]
      summary: Get a specific Kontablo account by ID
      operationId: accountsGet
      parameters:
        - name: account_id
          in: path
          required: true
          schema:
            type: string
          example: "asset.current.receivables"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KontabloAccount"
          description: Account details
        "404":
          $ref: "#/components/responses/NotFound"
...
  /classification/transaction:
    post:
      tags: [classification]
      summary: Classify a single journal entry transaction
      operationId: classifyTransaction
      description: |
        Given a transaction description (narration) + amount + context,
        returns the most probable Kontablo account mapping with confidence score.
        Uses multilingual semantic matching (supports 12 languages).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionClassificationRequest"
            example:
              narration: "Pago a proveedor Acme S.A. por materias primas"
              amount: 15000.00
              currency: "MXN"
              jurisdiction: "mx"
              debit_or_credit: "credit"
              context:
                industry: "manufacturing"
                company_size: "sme"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionClassificationResponse"
          description: Classification result
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    KontabloAccount:
      type: object
      properties:
        id:
          type: string
          example: "asset.current.receivables"
        uuid:
          type: string
          format: uuid
          example: "00000000-0000-4000-8000-000000000104"
        label_en:
          type: string
          example: "Trade Receivables"
        label_es:
          type: string
          example: "Cuentas por Cobrar Comerciales"
        ifrs_tag:
          type: string
          example: "ifrs-full:TradeAndOtherCurrentReceivables"
        nature:
          type: string
          enum: [debit, credit]
        statement:
          type: string
          enum: [balance_sheet, income_statement, cash_flow]
        level:
          type: integer
          enum: [1, 2, 3]
        parent:
          type: string
          nullable: true
          example: "asset.current"
        is_aggregate:
          type: boolean
        optional:
          type: boolean
          default: false
...
    TransactionClassificationRequest:
      type: object
      required: [narration, jurisdiction]
      properties:
        narration:
          type: string
          description: Transaction description / memo in any language
          maxLength: 500
        amount:
          type: number
          nullable: true
        currency:
          type: string
          nullable: true
        jurisdiction:
          type: string
        debit_or_credit:
          type: string
          enum: [debit, credit]
          nullable: true
        context:
          type: object
          properties:
            industry:
              type: string
            company_size:
              type: string
              enum: [micro, sme, large, enterprise]
    TransactionClassificationResponse:
      type: object
      properties:
        narration:
          type: string
        kontablo_id:
          type: string
        kontablo_uuid:
          type: string
          format: uuid
        label_en:
          type: string
        confidence_score:
          type: number
          format: float
        alternative_accounts:
          type: array
          items:
            type: object
            properties:
              kontablo_id:
                type: string
              confidence_score:
                type: number

    # ── Consolidation types ──
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        request_id:
          type: string